Improve exception messages on version conflicts (Issue #4524, PR #4540)
# Description

I updated the error message to try to give a better suggestion

I also refactored the logic to pull most of this into the exception class itself.

I didn't go for an explainer, because this logic can be contained in the exception and not much context is required.

closes #4524

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )