Fixed bug in eager promises on unset optional relations (PR #4534)
# Description

This PR moves evaluation of attribute references from `requires_emit` to `execute` (`ResultVariable.get_value()`) by introducing a result variable proxy. In addition to slightly improved semantics this allows for propagating unset variables. This is then used to fix the bug in the eager promises mechanism where a promise on an unset variable in a block that never gets executed would cause an exception. By propagating the unset variables, no execution units are left dangling (simply ignoring the exception would cause their requires to hang but by propagating the requires resolve to the unset variable one by one).

A couple of side effects:
- The `VariableReader` class has been greatly simplified. Some of its complexity has shifted to `IsDefinedGradualHelper`, the rest just ceased to be relevant.
- some error messages have slightly changed in the sense that failure now only occurs during execution rather than requires_emit. See tests changes for more details. I believe this is acceptable. I think the only fragile one is the `OptionalValueException` in `DynamicProxy` and that one has not changed (I've added a test for it).

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design