Support implicit removal of resource sets (Issue inmanta/inmanta-core#4529, PR #4555)
# Description

When a partial compile is ran, the exporter should trigger a deletion of each ResourceSet, defined in the partial model, that doesn't have any resources associated. In other words, the exporter should pass each ResourceSet without resources to the removed_resource_sets argument of the put_partial endpoint.

closes #4529

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [ ] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design