raises exception if removed resource sets are present in partial diff (Issue inmanta/inmanta-core#4455, PR #4597)
# Description

When a partial export is done with --remove-resource-sets myset but myset is still present in the resources that are exported, the server does not reject this. It should. This behavior should be well defined and documented.

Corner case: if the set does exist in the list but it is empty, there's no need to reject it.

closes https://github.com/inmanta/inmanta-core/issues/4455

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design