For v1tov2 conversion, split tag from version and put it in tag_build field (PR #4729)
# Description

When converting a v1 module to v2, if any version tags are present, don't put them in the version field but in the `tag_build` field. This prevents artifacts with versions like `2.11.1.dev0.dev202208182021`.

(In hindsight I probably should've ticketized this rather than to implement a fix, but what's done is done.)

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design