Fix race condition in test case `test_autostart_mapping_update_uri` (PR #4732)
# Description

The race condition is triggered by the API call to `client.set_setting()`. That endpoint starts a background task that call the `notify_agent_about_agent_map_update()` endpoint on the agent, which produces the log line of the form `Updating the URI of the endpoint <endpoint> from <x> to <y>` when the impacted agent has a session with the server. The race condition happens when the agent, started by the test case, establishes a session with the server before the background task started by the first call to `client.set_setting()` finishes. In that case the `notify_agent_about_agent_map_update()` endpoint is called twice on the agent before the first assertion on the log output happens. This results in the following two log lines:

```
Updating the URI of the endpoint internal from local: to <empty-string>
```
```
Updating the URI of the endpoint internal from <empty-string> to localhost
```
Instead of just one log line when the agent didn't establish a session with the server on time:

```
Updating the URI of the endpoint internal from local: to localhost
```

As such the test case fails, because it cannot find the expected log line in the log. I didn't see a reason for the first `client.set_setting()` call, so I removed it to prevent the race condition.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design