Improve query performance by splitting the query into two different queries. (PR #4841)
# Description

The original query first calculates the deploy status for all resources in the database before it actually filters. It might be possible to do this correct in a single query but it would make the query even more complex than it already is.

This PR does the following:

- fetch the resource without requires status
- do a second query that calculates the resource status only for the requires

The correct solution would be to normalize the data so that there is a table that contains the requires. This could be used to improve this query but can also replace the current provides field.

This solution is not perfect but it does reduce the query time with more than 10x for a serious size.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design