Fixed inconsistency and backwards compatibility for deprecated processing_events resource state (PR #4838)
# Description

Fixed inconsistency and backwards compatibility for deprecated processing_events resource state (introduced in #3020). We previously deleted the `processing_events` resource state because it was no longer required. We didn't delete it from the database however. This was a concious decision at the time to stay backwards compatible with older agents, but it turns out we overlooked some details: neither the API, nor the data model allowed the deprecated value anymore, causing any such values remaining in the database to result in errors.

This pull request removes the value from the database enum, and adds an explicit conversion to the `deploying` state in the entrypoint that might be called by older agents, for backwards compatibility.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design