Fixed two potential deadlocks with the cascading delete (PR #4889)
# Description

Fixes deadlock introduced by loosening explicit lock in #4871 without losing the huge performance gain that came with that PR.

Still in draft because there might be one more deadlock (a lot rarer). I would propose to merge the full PR before the release, except for `tests/test_postgres_concurrency.py`, which I'd like to do in a separate PR. If anyone thinks we should split it up further to have better confidence in the release, leave a comment about that and we'll discuss it.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design