Added guidelines for partial compiles (PR #4823)
# Description

Added modeling guidelines for safe use of partial compiles. I decided to put it in core rather than lsm because it could pretty much all be generalized and it all applies to core-only use of partial compiles as well.

Initial technical review. In the next stage I'll see if Pouriya is available for a review on the quality of the documentation.

- I had to move some files and GitHub does not show the diff properly. Review is probably easiest with `git diff master --color-moved` or `git diff master --find-renames=5%`

closes inmanta/lsm#482

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design