Ensure that the interface of AgentCache is backwards compatible with the previous implementation, because pytest-inmanta depends on it. (PR #4943)
# Description

Ensure that the interface of AgentCache is backwards compatible with the previous implementation, because pytest-inmanta depends on it. #4932 introduced the braking change. #4932 didn't get merged into ISO4 due to a merge conflict. That's why this PR only acts on master and iso5. I will create a separate PR to apply #4932 and this PR iso4 as well.

Belongs to #4932

# Self Check:

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design