Fixes an issue about optional fields without default value not being populated correctly in DAO (PR #5031)
# Description

Optional fields which were not assigned None as default were causing issues:
when populating a DAO using `__process_kwargs`, if the value of an optional field was
`None` in the DB, the DAO would not get this field populated. This caused issues when trying
to access the field as it did not exist.
This PR fixes this issue.

While resolving this, an inconsistency between the DB and the objects caused some tests to fail:
the `repository` and `branch` fields of an environment that equals None in objects are stored as empty strings in the DB

closes https://github.com/inmanta/inmanta-lsm/issues/1078

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design