Fix bug in deactivation logic of the ReentrantVirtualEnv used by the snippetcompiler. (PR #5051)
# Description

This PR fixes a bug in the deactivation logic of the ReentrantVirtualEnv, which is used by the compiler fixture for the compiler venv. The bug triggers when the `snippetcompiler` is used for the first time without calling into the `setup_for_snippet` method. That way the compiler venv is never created, because the `use_virtual_env` method of the `ReentrantVirtualEnv` class is never called. At the end of the first test case, the deactivation logic of the `ReentrantVirtualEnv` incorrectly assumes that the venv was created/activated. When the snippetcompiler fixture is then used in a second test case, the venv wouldn't be present on disk, while the `ReentrantVirtualEnv` class assumes is already exists. The result is that the second test case fails with the following error when it tries to do `pip install` using that compiler venv:

```
FileNotFoundError: [Errno 2] No such file or directory: '/tmp/tmpeqw57m5s/bin/python'
```

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design