Add support for extensions to register their environment settings via the `register_environment_settings` method in the `extension.py` file of the extension. (Issue inmanta/irt#1366, PR #5042)
I have been struggling at bit to find an elegant solution for this problem here. With this draft PR I would like to get feedback on the approach being used before I polish it. The general idea is that extensions define their custom environment settings via the `register_environment_settings` in the `extensions.py` method of the extension (See: https://github.com/inmanta/inmanta-lsm/pull/1084/files for an example). Inmanta-sphinx can then use the bootloader to discover all environment settings defined by the different extensions (see: https://github.com/inmanta/inmanta-sphinx/pull/57/files).

The implementation in inmanta-core is written as such that we have the possibility to only discover environment settings via the bootloader. This is required because executing the `setup()` of the license extensions results in an Exception when no valid license is available.

I also tried the approach where settings are exposed using a python entrypoint. That also didn't work, because it causes issues with circular dependencies. The following flow happens in that case:

1. `inmanta.data` discovers the entrypoints and invokes the method in the extension to obtain the environment settings.
2. Returning the settings again requires an import of `inmanta.data` which triggers the circular import

closes inmanta/irt#1366

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design