Fix bug where the status endpoint can become non-responsive (PR #5053)
# Description

This PR prevents the status endpoint form hanging when the DB is in trouble.

I produces a fully redundant, local cache of the compiler queue size.
I'm in general not a fan of redundant data like this, but the only alternative I see is to bake in a timeout in the status endpoint so it only reports data is can collect .e.g `<10ms`

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~