adapt constraints to not conflict with those of std (PR #5071)
# Description

The std module has a constraint on Jinja that is not present on core. Since the compiler environment is activated in-process after this dependency is already loaded in sys.modules, core's imcompatible version is used in practice. This commit will
constrain Jinja to the same version as it is constrained to in the std module.

closes #5004

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design