Reverted version bump of `packaging` on EOS because it contains breaking changes (PR #5245)
# Description

Turns out I made a mistake setting up dependabot after the new release. It still allowed dependabot version bumps on iso5. Most bumps are patch bumps we can safely leave as is but this once was a major bump that drops `LegacyVersion`, which has a small chance of being end user visible. I think we'd better keep it out of iso5.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )