Fix race condition in the `resource_action_consistency_check` method. (PR #5342)
Fix race condition in the `resource_action_consistency_check` method. The assertion would fail when more resource actions are added in between the execution of both SELECT queries. This PR ensures that both queries are executed in a transaction with the `repeatable_read` isolation level to prevent this race condition.

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~