Added ADR about timezone semantics (PR #5335)
# Description

I decided to add an ADR about this because I was once again confused about the fine points of how we decided to handle timezones.

Additionally, the ADR contains a new decision regarding new API endpoints. Please pay special attention to that. I already violated it in inmanta/designs#113 (before I double checked the current semantics), so I'll need to make minor changes there if this is accepted.

I've added the three of you as decision drivers and therefore reviewers because you were the reviewers of the PR that implemented this back in the day. Feel free to opt out if you like (as long as you don't do so all three of you).

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] ~~Type annotations are present~~
- [x] ~~Code is clear and sufficiently documented~~
- [x] ~~No (preventable) type errors (check using make mypy or make mypy-diff)~~
- [x] ~~Sufficient test cases (reproduces the bug/tests the requested feature)~~
- [x] Correct, in line with design
- [x] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~