Fixes some flakiness in the environment metric tests (Issue #5359, PR #5379)
# Description

Currently all the test for metric collector do the following steps:

1. starts the server
2. flushes metrics manuall
3. assert that the metrics are collected
This is potentially flaky as the server will also flush at regular intervals. We can never be sure amount the amount of metrics that are present in the metrics tables of the database when the assertions are executed.

this PR disables the automatic flush done at regular intervals

closes #5359

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )