Make sure a datapoint is written to the database even if no records are present for the monitored element. (Issue #5358, PR #5374)
# Description

When the database doesn't contain any record for the monitored element, no datapoint is written to the database. For example, when a given environment does not have any configuration models, the ResourceCountMetricsCollector will not return any MetricValue for that environment, while it should actually return the value 0 instead.

Tasks:

 Fix this bug for all collectors in inmanta-core that suffer from this bug.

closes #5358

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )