Ensure that each test case stops the agents it started. (PR #5422)
# Description

I noticed that some of these tasks are related to the heartbeat of an agent. So I assume this issue is caused by the fact that an agent is not stopped properly. This PR ensure that each test case stops the agents it starts. I will close this ticket after this fix. If the issue pops up again, we can reopen the ticket and investigate this further.

Closes #5332

# Self Check:

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~