Changed the default value of `environment_agent_trigger_method` environment setting to `push_incremental_deploy` (PR #5498)
# Description

In a recent support call, I have noticed that de dashboard promote button by default performs a full deploy.
This is not compatible with the upgrade procedures described in the docs.

1. I updated this default.
2. I tried to improve the docs

It is strictly speaking a breaking change, but it should not affect many things in practice.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [ ] ~~Type annotations are present~~
- [x] Code is clear and sufficiently documented
- [X] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] ~~Sufficient test cases (reproduces the bug/tests the requested feature)~~
- [x] Correct, in line with design
- [X] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )