Old agents in the agent table are now cleaned up from the database. (Issue #5349, PR #5586)
# Description

The agent table in the database was never cleaned up properly. It contained agents that were present in a very old version of the configuration model, while the currently released version doesn't have any resources for it anymore. This PR makes surre that old agents are cleaned up.

An agent can be deleted iff:

1. it's not used by any version
2. it's not in the agent map
3. it's down

closes #5349

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )