Add the user management service (Issue #5310, PR #5310)
# Description

This is a proof of concept version of having a built-in user mangement service. This has been an itch I have been having for a few years. The goal was to check how easy it would be to add this, turns out not that difficult actually. There are still quite some open question:

- How do we configure this?
- How do we get the initial user in the database?
  - A default user is a big no no
  - Add a command that needs to run on the server
  - Add special mode when there are no users, the frontend will make them create one
- When this is merged, do we enable auth by default?

There are still many todo's of which many are actually a ticket on their own.

I also had make some changes to how we construct queries because user also exists in the non public schema.

**Goals**

- Cleanup
- Make sure all tests succeed
- Get it through review

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )