Remove the code in the AutostartedAgentManager that ensure that the internal agent is present in the autostarted_agent_map. This code that safely be removed because a validator was added on the update path of the `autostart_agent_map` in Apr 21, 2020. So this check was included in the initial release of ISO4. (PR #5741)
# Description

This PR removes the code from the AutostartedAgentManager that ensures that the internal agent is present in the `autostarted_agent_map`. The code is removed because it performs a read-modify-write operation in a non-atomic way, that causes a race condition when the tests perform a call to update the `autostart_agent_map` setting.

The code to ensure that internal agent is present in the `autostart_agent_map` was introduced in [this PR](https://github.com/inmanta/inmanta-core/pull/2103),  because in an old version of the server there wasn't a validator on the update path of the setting that verified this. A [validator](https://github.com/inmanta/inmanta-core/blob/db32611489e71e1064d1c60df7b963200fd31114/src/inmanta/data/__init__.py#L2280) that perform this check was introduced in Apr 21, 2020, so before the initial release of ISO4. As such, we can safely remove this code now.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)