The 'inmanta module release' command now outputs the release tag (Issue #5816, PR #5885)
# Description

When the inmanta module release command is executed, it doesn't produce any output. That way you have to execute a git log to figure out which Git tag was created. This information is required to push the tag later on using the git push origin <tag> command. Now the release command will output the tag it created.

closes #5816

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)