Fix issue where the `test_server_recompile` test case times out because the compile takes longer than 30 seconds. (PR #6068)
# Description

The `test_server_recompile`  test case has been very flaky for a while. In some situations compiles take 20 seconds and sometimes they take more than 30 seconds. This is caused by the way we compose the pip commands. We supply some constraint to pip to make sure we don't change inmanta packages. This is done based on the working set (See: https://github.com/inmanta/inmanta-core/blob/132f0a7739464409f833e18dc12ec61ac98ae6ea/src/inmanta/env.py#L246). Because this is a set, the order of these constraints is undeterministic. Depending on this order, pip needs more or less time to resolve the required version, which reflects on the compile time. For example:

* `python -m pip install --upgrade --upgrade-strategy eager Jinja2<4,>=3.1 pydantic<2,>=1.10 email_validator<3,>=1.3 inmanta-core==9.1.0.dev0` happens to be a lot slower than on a venv with inmanta installed than
* `python -m pip install --upgrade --upgrade-strategy eager email_validator<3,>=1.3 Jinja2<4,>=3.1 pydantic<2,>=1.10 inmanta-core==9.1.0.dev0`

This PR raises the timeout threshold from 30 to 40 second to make sure that the slow path doesn't result in a timeout anymore.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)